스크립트 나누기 및 주석 삭제
