--- static/my_page_style.css
@@ -1,64 +1,0 @@
-.my_bg{
-    color:black;
-    background-color: rgb(239, 247, 218);
-    text-align: center;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 60px;
-    margin-bottom: 28px;
-}
-.my_Jjim{
-    margin-right: 20px;
-}
-.my_info{
-    margin-left:20px;
-}
-.my_Jjim , .my_info{
-    background-color: white;
-    height: 40px;
-    width: 100px;
-    border-radius: 20px;
-}
-.my_Jjim a, .my_info a{
-    height: 40px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-#my_information{
-    display: none;  
-    margin-top: 10px;
-}
-.card{
-    margin-top: 10px;
-}
-
-.My_ID, .My_PW, .My_Nick{
-    display: flex;
-    width: 980px;
-    height: 200px;
-    margin: auto;
-    border:solid black;
-    border-width: 1px 0 1px 0;
-}   
-.ID, .PW, .Nick{
-    display: flex;
-    border: solid black;
-    border-width: 0 1px 0 0;
-    justify-content: center;
-    align-items: center;
-    width: 150px;
-    background-color: rgb(248, 245, 241);
-}
-.Nick_content, .ID_content, .PW_content{
-    display: flex;
-    justify-content: left;
-    align-items: center;
-    margin-left:20px
-}
-
-.show {
-    display: block;
-  }
