+++ templates/main.html
@@ -1,7 +1,8 @@
 {% extends "base.html" %}
 
 {% block head %}
-    <script defer src="{{ url_for('static', filename='main.js') }}"></script>
+    <script defer src="{{ url_for('static', filename='js/main.js') }}"></script>
+    <script defer src="{{ url_for('static', filename='js/submit.js') }}"></script>
 {% endblock %}
 
 {% block content %}
