+++ templates/myPage.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 
 {% block head %}
-    <script defer src="{{ url_for('static', filename='mypage.js') }}"></script>
+    <script defer src="{{ url_for('static', filename='js/mypage.js') }}"></script>
 {% endblock %}
 
 {% block content %}
