+++ templates/search.html
@@ -1,7 +1,8 @@
 {% extends "base.html" %}
 
 {% block head %}
-    <script defer src="{{ url_for('static', filename='search.js') }}"></script>
+    <script defer src="{{ url_for('static', filename='js/submit.js') }}"></script>
+    <script defer src="{{ url_for('static', filename='js/search.js') }}"></script>
 {% endblock %}
 
 {% block content %}
@@ -28,42 +29,80 @@
         </div>
     </div>
     <section class ="cardCloumn" id="card_list">
-        <div class ="cardRow2">
-            {% for text in search %}
-                <div class="ListBg">
-                    <div class="ListFlex">
-                        <div class="click-wrap" onclick="showDetail('{{text['id']}}', event)">
-                            <div class="imgHidden-box">
-                                <img src="{{text['imgsrc']}}" class="classImg">
-                                <button class="Option_Jjim">
-                                    <img src="../static/img/heart.svg" class="heart">
-                                </button>
-                            </div>  
-                            <h3 class ="title">{{text['title']}}</h3>
-                            <div class ="numCount">
-                                <div class = "likeNum">
-                                    <img src="../static/img/likeUp3.png">
-                                    <span>{{text['likes']}}</span>
+        <div class ="cardRow">
+            {% if R == True %}
+                {% for text in search_R %}
+                    <div class="ListBg">
+                        <div class="ListFlex">
+                            <div class="click-wrap" onclick="showDetail('{{text['id']}}', event)">
+                                <div class="imgHidden-box">
+                                    <img src="{{text['imgsrc']}}" class="classImg">
+                                    <button class="Option_Jjim">
+                                        <img src="../static/img/heart.svg" class="heart">
+                                    </button>
+                                </div>  
+                                <h3 class ="title">{{text['title']}}</h3>
+                                <div class ="numCount">
+                                    <div class = "likeNum">
+                                        <img src="../static/img/likeUp3.png">
+                                        <span>{{text['likes']}}</span>
+                                    </div>
+                                    <div class = "JjimNum">
+                                        <img src="../static/img/Rheart.svg">
+                                        <span>0</span>
+                                    </div>
                                 </div>
-                                <div class = "JjimNum">
-                                    <img src="../static/img/Rheart.svg">
-                                    <span>0</span>
+                                <hr/>
+                            </div>
+                            <div class="Option">
+                                <div class="Like Option_Like" onclick="updateLike('{{text['id']}}')">
+                                    <img src="../static/img/thumbsup.svg" class="thumbsUp">
+                                    <span class="likes-btn">추천하기</span>
                                 </div>
+                                <a class="Link" href="{{text['url']}}" target= '_blank'>
+                                    <span>바로가기</span>
+                                </a>
                             </div>
-                            <hr/>
-                        </div>
-                        <div class="Option">
-                            <div class="Like Option_Like" onclick="updateLike('{{text['id']}}')">
-                                <img src="../static/img/thumbsup.svg" class="thumbsUp">
-                                <span class="likes-btn">추천하기</span>
-                            </div>
-                            <a class="Link" href="{{text['url']}}" target= '_blank'>
-                                <span>바로가기</span>
-                            </a>
                         </div>
                     </div>
-                </div>
-            {% endfor %}
+                {% endfor %}
+            {% elif L == True %}
+                {% for text in search_L %}
+                        <div class="ListBg">
+                            <div class="ListFlex">
+                                <div class="click-wrap" onclick="showDetail('{{text['id']}}', event)">
+                                    <div class="imgHidden-box">
+                                        <img src="{{text['imgsrc']}}" class="classImg">
+                                        <button class="Option_Jjim">
+                                            <img src="../static/img/heart.svg" class="heart">
+                                        </button>
+                                    </div>  
+                                    <h3 class ="title">{{text['title']}}</h3>
+                                    <div class ="numCount">
+                                        <div class = "likeNum">
+                                            <img src="../static/img/likeUp3.png">
+                                            <span>{{text['likes']}}</span>
+                                        </div>
+                                        <div class = "JjimNum">
+                                            <img src="../static/img/Rheart.svg">
+                                            <span>0</span>
+                                        </div>
+                                    </div>
+                                    <hr/>
+                                </div>
+                                <div class="Option">
+                                    <div class="Like Option_Like" onclick="updateLike('{{text['id']}}')">
+                                        <img src="../static/img/thumbsup.svg" class="thumbsUp">
+                                        <span class="likes-btn">추천하기</span>
+                                    </div>
+                                    <a class="Link" href="{{text['url']}}" target= '_blank'>
+                                        <span>바로가기</span>
+                                    </a>
+                                </div>
+                            </div>
+                        </div>
+                {% endfor %}
+            {% endif %}
         </div>     
     </section>
     <!-- 리스트 상세보기 -->
